Simplifying the backend, transitioning from PostgreSQL to SQLite
