+++ backend/src/app.module.ts
@@ -1,47 +1,22 @@
 import { Module } from '@nestjs/common';
-import { ConfigModule, ConfigService } from '@nestjs/config';
+import { ConfigModule } from '@nestjs/config';
 import { APP_GUARD } from '@nestjs/core';
-import { JwtModule } from '@nestjs/jwt';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { AuthModule } from './auth/auth.module';
-import jwtConfig from './auth/config/jwt.config';
-import { AccessTokenGuard } from './auth/guards/access-token.guard';
 import { AuthGuard } from './auth/guards/auth.guard';
-import databaseConfig from './config/database.config';
-import environmentValidation from './config/environment.validation';
+import { appConfig } from './config/app.config';
+import { dbConfig } from './config/db.config';
 import { ProfilesModule } from './profiles/profiles.module';
 import { UsersModule } from './users/users.module';
 
-const ENV = process.env.NODE_ENV;
+const modules = [AuthModule, UsersModule, ProfilesModule];
 
 @Module({
   imports: [
-    AuthModule,
-    UsersModule,
-    ProfilesModule,
-    ConfigModule.forFeature(jwtConfig),
-    JwtModule.registerAsync(jwtConfig.asProvider()),
-    ConfigModule.forRoot({
-      isGlobal: true,
-      envFilePath: ENV ? `.env.${ENV}` : '.env',
-      load: [databaseConfig],
-      validationSchema: environmentValidation,
-    }),
-    TypeOrmModule.forRootAsync({
-      imports: [ConfigModule],
-      inject: [ConfigService],
-      useFactory: (configService: ConfigService) => ({
-        type: 'postgres',
-        host: configService.get('database.host'),
-        port: configService.get('database.port'),
-        username: configService.get('database.username'),
-        password: configService.get('database.password'),
-        database: configService.get('database.name'),
-        synchronize: configService.get('database.synchronize'),
-        autoLoadEntities: configService.get('database.autoLoadEntities'),
-      }),
-    }),
+    ...modules,
+    ConfigModule.forRoot(appConfig),
+    TypeOrmModule.forRoot(dbConfig),
   ],
-  providers: [{ provide: APP_GUARD, useClass: AuthGuard }, AccessTokenGuard],
+  providers: [{ provide: APP_GUARD, useClass: AuthGuard }],
 })
 export class AppModule {}
