+++ backend/src/auth/constants/auth.constants.ts
@@ -1,2 +1,2 @@
 export const REQUEST_USER_KEY = 'user';
-export const AUTH_TYPE_KEY = 'authType';
+export const PUBLIC_KEY = 'public';
