+++ backend/src/auth/decorators/public.decorator.ts
@@ -1,0 +1,4 @@
+import { SetMetadata } from '@nestjs/common';
+import { PUBLIC_KEY } from '../constants/auth.constants';
+
+export const Public = () => SetMetadata(PUBLIC_KEY, true);
