--- backend/src/auth/enums/auth-type.enum.ts
@@ -1,4 +1,0 @@
-export enum AuthType {
-  Bearer,
-  None,
-}
