--- backend/src/auth/guards/access-token.guard.ts
@@ -1,51 +1,0 @@
-import {
-  CanActivate,
-  ExecutionContext,
-  Inject,
-  Injectable,
-  UnauthorizedException,
-} from '@nestjs/common';
-import { ConfigType } from '@nestjs/config';
-import { JwtService } from '@nestjs/jwt';
-import jwtConfig from '../config/jwt.config';
-import { REQUEST_USER_KEY } from '../constants/auth.constants';
-import { Request } from 'express';
-
-@Injectable()
-export class AccessTokenGuard implements CanActivate {
-  constructor(
-    @Inject(jwtConfig.KEY)
-    private readonly jwtConfiguration: ConfigType<typeof jwtConfig>,
-    private readonly jwtService: JwtService,
-  ) {}
-
-  async canActivate(context: ExecutionContext): Promise<boolean> {
-    const request = context.switchToHttp().getRequest<Request>();
-    const accessToken = request.headers.authorization?.split(' ')[1];
-
-    if (!accessToken) {
-      throw new UnauthorizedException(
-        'You are not authorized, the access token is missing.',
-      );
-    }
-
-    try {
-      const payload = await this.jwtService.verifyAsync(accessToken, {
-        ...this.jwtConfiguration,
-        ignoreExpiration: false,
-      });
-
-      request[REQUEST_USER_KEY] = payload;
-
-      return true;
-    } catch (error) {
-      if (error?.name === 'TokenExpiredError') {
-        throw new UnauthorizedException('Access token has expired');
-      }
-      if (error?.name === 'JsonWebTokenError') {
-        throw new UnauthorizedException('Invalid access token');
-      }
-      throw new UnauthorizedException('Authentication failed');
-    }
-  }
-}
