--- backend/src/auth/providers/bcrypt.provider.ts
@@ -1,18 +1,0 @@
-import { Injectable } from '@nestjs/common';
-import * as bcrypt from 'bcrypt';
-import { HashingProvider } from './hashing.provider';
-
-@Injectable()
-export class BcryptProvider implements HashingProvider {
-  public async hashPassword(data: string): Promise<string> {
-    const salt = await bcrypt.genSalt();
-    return bcrypt.hash(data, salt);
-  }
-
-  public async comparePassword(
-    data: string,
-    encrypted: string,
-  ): Promise<boolean> {
-    return bcrypt.compare(data, encrypted);
-  }
-}
