+++ backend/src/auth/providers/hashing.provider.ts
@@ -1,7 +1,16 @@
 import { Injectable } from '@nestjs/common';
+import * as bcrypt from 'bcrypt';
 
 @Injectable()
-export abstract class HashingProvider {
-  abstract hashPassword(data: string): Promise<string>;
-  abstract comparePassword(data: string, encrypted: string): Promise<boolean>;
+export class HashingProvider {
+  public async hashPassword(data: string): Promise<string> {
+    const salt = await bcrypt.genSalt();
+    return bcrypt.hash(data, salt);
+  }
+  public async comparePassword(
+    data: string,
+    encrypted: string,
+  ): Promise<boolean> {
+    return bcrypt.compare(data, encrypted);
+  }
 }
