--- backend/src/auth/token-blacklist.entity.ts
@@ -1,21 +1,0 @@
-import {
-  Column,
-  CreateDateColumn,
-  Entity,
-  PrimaryGeneratedColumn,
-} from 'typeorm';
-
-@Entity()
-export class TokenBlacklist {
-  @PrimaryGeneratedColumn()
-  id: number;
-
-  @Column()
-  token: string;
-
-  @Column()
-  expiresAt: Date;
-
-  @CreateDateColumn()
-  createdAt: Date;
-}
