--- backend/src/config/database.config.ts
@@ -1,11 +1,0 @@
-import { registerAs } from '@nestjs/config';
-
-export default registerAs('database', () => ({
-  host: process.env.DATABASE_HOST,
-  port: parseInt(process.env.DATABASE_PORT),
-  username: process.env.DATABASE_USERNAME,
-  password: process.env.DATABASE_PASSWORD,
-  name: process.env.DATABASE_NAME,
-  synchronize: process.env.DATABASE_SYNC === 'true' ? true : false,
-  autoLoadEntities: process.env.DATABASE_AUTO_LOAD === 'true' ? true : false,
-}));
