+++ backend/src/config/env.validation.ts
@@ -1,0 +1,12 @@
+import * as Joi from 'joi';
+
+export default Joi.object({
+  NODE_ENV: Joi.string()
+    .valid('development', 'production')
+    .default('development'),
+  JWT_SECRET: Joi.string().required(),
+  JWT_TOKEN_AUDIENCE: Joi.string().required(),
+  JWT_TOKEN_ISSUER: Joi.string().required(),
+  JWT_ACCESS_TOKEN_TTL: Joi.number().required(),
+  JWT_REFRESH_TOKEN_TTL: Joi.number().required(),
+});
