--- backend/src/config/environment.validation.ts
@@ -1,28 +1,0 @@
-import * as Joi from 'joi';
-
-export default Joi.object({
-  NODE_ENV: Joi.string()
-    .valid('development', 'production', 'test')
-    .default('development'),
-  // DATABASE
-  DATABASE_HOST: Joi.string().required(),
-  DATABASE_PORT: Joi.number().port().default(5432),
-  DATABASE_USERNAME: Joi.string().required(),
-  DATABASE_PASSWORD: Joi.string().required(),
-  DATABASE_NAME: Joi.string().required(),
-  DATABASE_SYNC: Joi.string().optional(),
-  DATABASE_AUTO_LOAD: Joi.string().optional(),
-  // JWT AUTH
-  JWT_SECRET: Joi.string().required(),
-  JWT_TOKEN_AUDIENCE: Joi.string().required(),
-  JWT_TOKEN_ISSUER: Joi.string().required(),
-  JWT_ACCESS_TOKEN_TTL: Joi.number().required(),
-  JWT_REFRESH_TOKEN_TTL: Joi.number().required(),
-
-  // AWS S3 Bucket
-  AWS_PUBLIC_BUCKET_NAME: Joi.string().required(),
-  AWS_REGION: Joi.string().required(),
-  AWS_CLOUDFRONT_URL: Joi.string().required(),
-  AWS_ACCESS_KEY_ID: Joi.string().required(),
-  AWS_SECRET_ACCESS_KEY: Joi.string().required(),
-});
